+++ CHANGES.md
@@ -1,3 +1,9 @@
+# 20.11
+
+   * The aprox19 + NSE network was ported to C++ (#362)
+
+   * The simplified-SDC code path was ported to C++ (#389)
+
 # 20.10
 
    * An option to use NSE instead of integrating the reaction